ServiceRaiting

+++ Cinema.Movie/Cinema.Movie.Web/Modules/Common/Migrations/DefaultDB/DefaultDB_20160607_115100_Initial.cs
@@ -105,7 +105,7 @@ namespace Cinema.Movie.Migrations.DefaultDB
 
             Create.Table("ServiceRating").InSchema("mov")
                .WithColumn("ServiceRatingId").AsInt16().Identity().PrimaryKey().NotNullable()
-               .WithColumn("Rating").AsInt16().NotNullable().WithDefaultValue(0)
+               .WithColumn("Rating").AsDouble().NotNullable().WithDefaultValue(0)
                .WithColumn("MovieId").AsInt64().NotNullable()
                     .ForeignKey("FK_ServiceRating_MovieId", "mov", "Movie", "MovieId")
                .WithColumn("ServiceId").AsInt16()
